refactor: 테크블로그 폴더구조 개선
