+++ src/pages/tech_blog/style/PostCard.style.ts
